Refactor

+++ src/resolvers/author.resolver.ts
@@ -98,7 +98,14 @@ export class AuthorResolver {
     async deleteOneAuthor(
         @Arg("input", () => AuthorIdInput) input: AuthorIdInput
     ): Promise<Boolean> {
-        await this.authorRepository.delete(input.id);
-        return true;
+        try {
+            const author = await this.authorRepository.findOne(input.id);
+            if (!author) throw new Error('Author does not exist');
+            await this.authorRepository.delete(input.id);
+            return true;
+        } catch (e) {
+            throw new Error(e.message)
+        }
+
     }
 }
